refactor(slm-frontend): consolidate duplicated axios-compat adapter into slmApiCompat (#12654)
