+++ autobot-slm-frontend/src/composables/useRoles.test.ts
@@ -26,7 +26,7 @@
  */
 
 import { describe, it, expect, vi, beforeEach } from 'vitest'
-import { jsonResponse, errorResponse } from './useSlmApi.testHelper'
+import { jsonResponse, errorResponse } from './slmApiClient.testHelper'
 
 const mockRaw = vi.fn()
