+++ autobot-slm-frontend/src/composables/useSlmApi.ts
@@ -9,7 +9,7 @@
  * Provides REST API integration for all SLM endpoints.
  */
 
-import { slmApiClient } from '@/utils/ApiClient'
+import { makeAxiosCompatClient } from '@/utils/slmApiCompat'
 import type {
   SLMNode,
   NodeHealth,
@@ -98,94 +98,16 @@ import type {
 } from '@/types/api-responses'
 
 // =============================================================================
-// slmApiClient adapter (#12420 Phase 2 batch 5)
+// slmApiClient adapter (#12420 Phase 2 batch 5 → consolidated in #12654)
 //
-// useSlmApi historically owned its own axios instance (base URL from
-// getSlmApiBase(), a request interceptor injecting the SLM bearer token, and no
-// response interceptor). This adapter routes every call through the canonical
-// `slmApiClient` — where the auth token, base URL and centralised 401 handling
-// now live — while reproducing the axios surface the methods below depend on so
-// their bodies, and all ~26 consumers, stay byte-for-byte unchanged:
-//
-//   * methods `await client.<verb>(endpoint, body?)` and read `response.data`
-//     → the adapter returns `{ data }`.
-//   * consumers read `err.response.status` / `err.response.data.detail` on a
-//     rejected call (e.g. upsertSecret's 409 fallback, SetupWizardView,
-//     SecretsSettings) → the adapter throws an axios-shaped error carrying
-//     `response.status` and `response.data`.
-//
-// It delegates to `slmApiClient.rawRequest` (not the get/post/... helpers) on
-// purpose: rawRequest is the single seam that injects the bearer token + base
-// URL and runs the 401 handler, WITHOUT the helpers' GET retry/back-off or the
-// `HTTP <n>: <msg>` error transform — preserving the original single-shot,
-// structured-error behaviour the consumers rely on.
+// useSlmApi historically owned its own axios instance; the Phase-2 migration
+// routed every call through the canonical `slmApiClient` via a thin axios-compat
+// adapter. That adapter now lives in `utils/slmApiCompat.ts` (shared with
+// useRoles). `textFallback: true` preserves this composable's behaviour of
+// returning `response.text()` for a non-JSON 2xx body (PEM cert downloads).
 // =============================================================================
 
-interface AxiosLikeResponse<T> {
-  data: T
-}
-
-// Serialise an axios-style params object onto the endpoint. Only getProvisionStatus
-// passes a params object; every other method already builds its own query string.
-function withParams(endpoint: string, params?: Record<string, unknown>): string {
-  if (!params) return endpoint
-  const usp = new URLSearchParams()
-  for (const [key, value] of Object.entries(params)) {
-    if (value !== undefined && value !== null) usp.append(key, String(value))
-  }
-  const qs = usp.toString()
-  if (!qs) return endpoint
-  return endpoint.includes('?') ? `${endpoint}&${qs}` : `${endpoint}?${qs}`
-}
-
-async function adapterRequest<T>(
-  method: string,
-  endpoint: string,
-  body?: unknown
-): Promise<AxiosLikeResponse<T>> {
-  const response = await slmApiClient.rawRequest(endpoint, { method, body })
-
-  if (!response.ok) {
-    let data: unknown = null
-    try {
-      data = await response.json()
-    } catch {
-      /* non-JSON error body — leave data null, mirroring axios */
-    }
-    const error = new Error(`HTTP ${response.status}`) as Error & {
-      response: { status: number; data: unknown }
-    }
-    // Reproduce the axios error shape consumers read (err.response.status/.data).
-    error.response = { status: response.status, data }
-    throw error
-  }
-
-  if (response.status === 204) return { data: {} as T }
-  const contentType = response.headers.get('content-type')
-  if (contentType && contentType.includes('application/json')) {
-    return { data: (await response.json()) as T }
-  }
-  // Non-JSON 2xx (e.g. PEM cert downloads) — return the raw text, as axios does
-  // when its default JSON transform cannot parse the payload.
-  return { data: (await response.text()) as unknown as T }
-}
-
-// Axios-compatible facade over slmApiClient consumed by every method below.
-const client = {
-  get: <T = unknown>(
-    endpoint: string,
-    config?: { params?: Record<string, unknown> }
-  ): Promise<AxiosLikeResponse<T>> =>
-    adapterRequest<T>('GET', withParams(endpoint, config?.params)),
-  post: <T = unknown>(endpoint: string, body?: unknown): Promise<AxiosLikeResponse<T>> =>
-    adapterRequest<T>('POST', endpoint, body),
-  put: <T = unknown>(endpoint: string, body?: unknown): Promise<AxiosLikeResponse<T>> =>
-    adapterRequest<T>('PUT', endpoint, body),
-  patch: <T = unknown>(endpoint: string, body?: unknown): Promise<AxiosLikeResponse<T>> =>
-    adapterRequest<T>('PATCH', endpoint, body),
-  delete: <T = unknown>(endpoint: string): Promise<AxiosLikeResponse<T>> =>
-    adapterRequest<T>('DELETE', endpoint),
-}
+const client = makeAxiosCompatClient({ textFallback: true })
 
 // Backend response types (different from frontend SLMNode)
 interface BackendNodeResponse {
